Improve /tempos command and refactor streaming commands
There was no need for the /tempos stream to redundantly include the
master tempo information, and it became clear that streaming commands
were mostly identical to each other, so I was able to refactor them to
use a shared function with configuration arguments, removing a bunch
of code. This made it easy to also improve the /master command so that
it could be queried for current state as well.